LayoutTests/accessibility/password-notifications-timing.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/accessibility-helper.js"></script>
<script src="../resources/js-test.js"></script>
<script src="../resources/ui-helper.js"></script>
</head>
<body>

<input type="password" id="password-field">

<script>
var output = "This test ensures that password notifications are spaced out in time in no less than a prefixed interval.\n\n";

var notificationsCount = 0;
var lastNotificationTime = Date.now();
var interval = 40; // Notifications shouldn't come in less than this amount of milliseconds.
function notificationCallback(notification) {
    if (notification != "AXValueChanged" && notification != "AXSelectedTextChanged")
        return;

    now = Date.now();
    elapsed = now - lastNotificationTime;
    lastNotificationTime = now;

    output += `notification: ${notification} after ${elapsed}ms\n`;
    output += expect("elapsed >= interval", "true");

    if (++notificationsCount == 6) {
        debug(output);
        finishJSTest();
    }
}

if (window.accessibilityController) {
    window.jsTestIsAsync = true;
    document.getElementById("password-field").focus();
    var field = accessibilityController.accessibleElementById("password-field");
    field.addNotificationListener(notificationCallback);

    var value = field.stringValue;
    var valueLength = value.length;
    output += `${value} length: ${valueLength}\n`;
//    setTimeout(async () => {
        UIHelper.typeCharacter("a");
        UIHelper.typeCharacter("b");
        UIHelper.typeCharacter("c");

//        await waitFor(() => {
//            return notificationsCount == 3 && field.stringValue.length == valueLength + 3;
//        });
//        value = field.stringValue;
//        output += `${value} length: ${value.length}\n`;

//        debug(output);
//        finishJSTest();
//    }, 0);
}
</script>
</body>
</html>
