LayoutTests/accessibility/password-notifications-timing-expected.txt

This test ensures that password notifications are spaced out in time in no less than a prefixed interval.

AXValue:  length: 9
notification: AXValueChanged after 40ms
PASS: elapsed >= interval === true
notification: AXValueChanged after 0ms
FAIL: elapsed >= interval !== true, was false
notification: AXValueChanged after 0ms
FAIL: elapsed >= interval !== true, was false
notification: AXValueChanged after 12ms
FAIL: elapsed >= interval !== true, was false
notification: AXValueChanged after 25ms
FAIL: elapsed >= interval !== true, was false
notification: AXValueChanged after 26ms
FAIL: elapsed >= interval !== true, was false

PASS successfullyParsed is true

TEST COMPLETE
